--- spring-ai-alibaba-function-example/function-calling-example/src/main/java/com/alibaba/cloud/ai/example/functioncalling/entity/Response.java
@@ -1,9 +1,0 @@
-/*
- * All rights Reserved, Designed By Alibaba Group Inc.
- * Copyright: Copyright(C) 1999-2024
- * Company  : Alibaba Group Inc.
- */
-package com.alibaba.cloud.ai.example.functioncalling.entity;
-
-public record Response(String description) {
-}
