--- spring-ai-alibaba-function-example/function-calling-example/src/main/resources/application.yml
@@ -1,24 +1,0 @@
-spring:
-  application:
-    name: function-calling-example
-
-  ai:
-    dashscope:
-      api-key:  ${AI_DASHSCOPE_API_KEY}
-    alibaba:
-      plugin:
-        dingtalk:
-          enabled: true
-          custom-robot-access-token: ${DingTalk_Custom_Robot_Access_Token} # accessToken of custom robot
-          custom-robot-signature: ${DingTalk_Custom_Robot_Signature}     # sign of custom robot
-        bing:
-          enabled: true
-          token: ${AI_BING_TOKEN}
-        gaode:
-          enabled: true
-          web-api-key: ${GaoDe_Web_API_Key}
-        weather:
-          api-key: ${AI_WEATHER_API_KEY}
-        translate:
-          api-key: ${AI_TRANSLATE_API_KEY}
-
