+++ spring-ai-alibaba-function-example/src/main/java/com/alibaba/cloud/ai/functioncalling/MainApplication.java
@@ -14,7 +14,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package com.alibaba.cloud.ai.functioncalling;
 
 import org.springframework.boot.SpringApplication;
