+++ spring-ai-alibaba-function-example/src/main/resources/application.yml
